# AnotherMe/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnotherMe Gateway API",
    "version": "0.3.0",
    "description": "Unified backend gateway for AnotherMe Web and Mobile clients.\n\nCapabilities follow the registry defined in `agents/foundation/capability_registry.py`.\nValid capability ids: `course_generate`, `problem_video_generate`, `quiz_practice`,\n`interactive_demo`, `ai_tutor_chat`, `deep_solve`, `deep_research`,\n`math_animator`, `visualize`, `co_writer`.\n\nP1 changes:\n  - Capability endpoints (P0-3) are now wired.\n  - Global error envelope (`{ error_code, message, request_id, details? }`).\n  - Live-book / co-writer / courses routes are first-class.\n\nAuth: Phase 0 uses static `Bearer <GATEWAY_API_TOKEN>`. Phase 2 will switch to Supabase JWT.",
    "contact": {
      "name": "AnotherMe",
      "url": "https://github.com/anotherme/anotherme"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Phase 0: static token from GATEWAY_API_TOKEN env var. Phase 2: Supabase JWT."
      }
    },
    "schemas": {
...
      "ClassroomSummary": {
        "type": "object",
        "required": ["id", "title", "created_at", "scenes_count"],
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "created_at": { "type": "string" },
          "scenes_count": { "type": "integer" }
        }
      },
      "ClassroomListResponse": {
        "type": "object",
        "required": ["success", "classrooms"],
        "properties": {
          "success": { "type": "boolean" },
          "classrooms": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ClassroomSummary" }
          }
        }
      },
      "ClassroomData": {
        "type": "object",
        "additionalProperties": true
      },
      "ClassroomDetailResponse": {
        "type": "object",
        "required": ["success", "classroom"],
        "properties": {
          "success": { "type": "boolean" },
          "classroom": { "$ref": "#/components/schemas/ClassroomData" }
        }
      },
      "CreateClassroomRequest": {
        "type": "object",
        "required": ["stage"],
        "properties": {
          "stage": { "type": "object", "additionalProperties": true },
          "scenes": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
      "CreateClassroomResponse": {
        "type": "object",
        "required": ["success", "id"],
        "properties": {
          "success": { "type": "boolean" },
          "id": { "type": "string" }
        }
      },
      "DeleteClassroomResponse": {
        "type": "object",
        "required": ["success", "deleted"],
        "properties": {
          "success": { "type": "boolean" },
          "deleted": { "type": "boolean" }
        }
      },
...
      "StudentProfileOutput": {
        "type": "object",
        "required": ["user_id", "learning_stats", "computed_at", "profile_source"],
        "properties": {
          "user_id": { "type": "string" },
          "weak_subjects": { "type": "array", "items": { "type": "string" } },
          "weak_knowledge_points": { "type": "array", "items": { "type": "string" } },
          "recent_focus": { "type": "string" },
          "ability_scores": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "learning_stats": { "type": "object", "additionalProperties": true },
          "updated_at": { "type": "string" },
          "computed_at": { "type": "string" },
          "profile_source": { "type": "string" },
          "knowledge_tracing": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
...
      "APIError": {
        "type": "object",
        "required": ["error_code", "message", "request_id"],
        "properties": {
          "error_code": {
            "type": "string",
            "example": "INTERNAL_ERROR",
            "description": "Machine-readable error code (UPPER_SNAKE_CASE)."
          },
          "message": { "type": "string", "example": "Internal server error" },
          "request_id": { "type": "string", "description": "Echoed from X-Request-Id header, or auto-minted." },
          "details": { "type": "object", "additionalProperties": true }
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/APIError" }
          }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/students/{user_id}/profile": {
      "get": {
        "tags": ["knowledge"],
        "summary": "Get a student's learning profile",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Student profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentProfileOutput" }
              }
            }
          }
        }
      }
    },
...
    "/v1/classrooms": {
      "get": {
        "tags": ["classroom"],
        "summary": "List classrooms (most recent first)",
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "Classrooms",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClassroomListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["classroom"],
        "summary": "Create a classroom (called by Web app via Next.js)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateClassroomRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Classroom created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateClassroomResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/classrooms/{classroom_id}": {
      "get": {
        "tags": ["classroom"],
        "summary": "Get a classroom (mobile entry-point)",
        "parameters": [
          { "name": "classroom_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Classroom data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClassroomDetailResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["classroom"],
        "summary": "Delete a classroom",
        "parameters": [
          { "name": "classroom_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Classroom deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteClassroomResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  }
}